Extract WorkerSource types to separate file
